requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1        
asn1crypto==1.5.1     
beautifulsoup4==4.14.2
cachetools==6.2.2     
certifi==2025.11.12   
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
cssselect2==0.8.0
Django==4.2.26
django-cors-headers==4.9.0
django-environ==0.11.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
environ==1.0
et_xmlfile==2.0.0
google==3.0.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
html5lib==1.1
httplib2==0.31.0
idna==3.10
lxml==5.3.0
mysql-connector-python==9.1.0
mysqlclient==2.2.7
numpy==2.0.2
oauthlib==3.3.1
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.3.3
paywix==2.0.2
pillow==11.0.0
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==5.1.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.16.0
soupsieve==2.8
sqlparse==0.5.2
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.2
uritemplate==4.2.0
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
xhtml2pdf==0.2.16
xlwt==1.3.0